requirements.txt
absl-py==0.9.0
appdirs==1.4.4
APScheduler==3.6.3
astor==0.8.1
astunparse==1.6.3
attrs==19.3.0
bcrypt==3.1.7
botocore==1.17.43
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cliff==3.4.0
cmd2==1.3.3
colorama==0.4.3
cryptography==3.0
cycler==0.10.0
debtcollector==2.2.0
decorator==4.4.2
dnspython==1.16.0
docutils==0.15.2
dogpile.cache==1.0.2
eventlet==0.26.1
extras==1.0.0
face-recognition-models==0.3.0
fasteners==0.15
fixtures==3.0.0
Flask==1.1.2
Frozen-Flask==0.15
gast==0.3.3
google-auth==1.19.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.16
grpcio==1.30.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
imutils==0.5.3
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
jsonpatch==1.26
jsonpointer==2.0
jsonschema==3.2.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keystoneauth1==4.2.1
kiwisolver==1.2.0
linecache2==1.0.0
mahotas==1.4.10
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.0
monotonic==1.5
msgpack==1.0.0
mtcnn==0.1.0
munch==2.5.0
myapp==0.1.dev0
netaddr==0.8.0
netifaces==0.10.9
numpy==1.18.5
oauthlib==3.1.0
opencv-contrib-python==4.3.0.36
opencv-python==4.3.0.36
openstacksdk==0.48.0
opt-einsum==3.3.0
os-client-config==2.1.0
os-service-types==1.7.0
osc-lib==2.2.0
oslo.concurrency==4.2.0
oslo.config==8.3.1
oslo.context==3.1.0
oslo.i18n==5.0.0
oslo.log==4.3.0
oslo.serialization==4.0.0
oslo.utils==4.4.0
packaging==20.4
pandas==1.1.0
paramiko==2.7.1
Paste==3.4.3
PasteDeploy==2.1.0
pbr==5.4.5
Pillow==7.2.0
prettytable==0.7.2
protobuf==3.12.2
psutil==5.7.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.0
pyreadline==2.1
pyrsistent==0.16.0
python-cinderclient==7.1.0
python-dateutil==2.8.1
python-freezerclient==3.1.0
python-glanceclient==3.2.1
python-keystoneclient==4.1.0
python-mimeparse==1.6.0
python-neutronclient==7.2.0
python-novaclient==17.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
requestsexceptions==1.4.0
rfc3986==1.4.0
rsa==4.6
scikit-learn==0.23.1
scipy==1.4.1
simplejson==3.17.2
six==1.15.0
sklearn==0.0
stevedore==3.2.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-cpu==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
testtools==2.4.0
threadpoolctl==2.1.0
traceback2==1.4.0
tzlocal==2.1
unittest2==1.1.0
urllib3==1.25.10
warlock==1.3.3
wcwidth==0.2.5
WebOb==1.8.6
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
xgboost==1.1.1
zipp==3.1.0
